% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_nrep.R
\name{simulate_nrep}
\alias{simulate_nrep}
\title{Title}
\usage{
simulate_nrep(n_rep, n_samp, beta0, beta1)
}
\arguments{
\item{beta1}{}
}
\description{
Title
}
\examples{
# simulate a single dataset
sim_regression(30, 2, 3)
}
